Moved major page components to the top level folder of the theme
