+++ pages/signin.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import SignIn from "../.current_theme/components/SignIn";
+import SignIn from "../.current_theme/SignIn";
 export default {
     components: {
         SignIn,
