+++ pages/signup.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import SignUp from "../.current_theme/components/SignUp";
+import SignUp from "../.current_theme/SignUp";
 export default {
     components: {
         SignUp,
