+++ themes/damal/SignUp.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import SignupForm from "./SignupForm";
+import SignupForm from "./components/SignupForm";
 export default {
     name: "SignUp",
     components: {
